refactor: 커리큘럼 페이지 수정사항 반영 develop 브랜치에 있는 커리큘럼 페이지명으로 변경
